refactor/demo project for better demonstration
refactor/demo, paths and project structure \n\n del/pure css reference \n\n feat/material design 3 \n\n feat/playground and get started section
