+++ demo/index.html
@@ -1,77 +1,90 @@
 <!doctype html>
 <html lang="en">
   <head>
-    <title>Codemirror Themes Example</title>
+    <meta charset="UTF-8" />
+    <title>&#64;fsegurai/codemirror-theme | Demo</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="icon" type="image/x-icon" href="public/favicon.ico" />
+
+    <!-- Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       rel="stylesheet"
-      href="https://unpkg.com/purecss@2.0.6/build/pure-min.css"
-      integrity="sha384-Uu6IeWbM+gzNVXJcM9XV3SohHtmWE+3VGi496jvgX1jyvDTXfdK+rfZc8C1Aehk5"
-      crossorigin="anonymous"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined&family=Open+Sans:ital,wght@0,300..800;1,300..800&display=swap"
     />
-    <link rel="icon" type="image/x-icon" href="favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <script async type="module" src="./dist/index.js"></script>
-    <style>
-      * {
-        -webkit-box-sizing: border-box;
-        -moz-box-sizing: border-box;
-        box-sizing: border-box;
-      }
 
-      .home-menu {
-        background: #2d3e50;
-      }
+    <!-- Local Styles -->
+    <link rel="stylesheet" href="./styles/themes.css" />
+    <link rel="stylesheet" href="./styles/header.css" />
+    <link rel="stylesheet" href="./styles/index.css" />
 
-      .home-menu .pure-menu-heading {
-        color: white;
-        font-weight: 400;
-        font-size: 120%;
+    <!-- Import Map for @material/web -->
+    <script type="importmap">
+      {
+        "imports": {
+          "@material/web/": "https://esm.run/@material/web/"
+        }
       }
+    </script>
 
-      .home-menu .pure-menu-selected a.pure-menu-link {
-        color: white;
-      }
+    <!-- Material Web Components -->
+    <script type="module">
+      import '@material/web/all.js'
+    </script>
 
-      .home-menu a {
-        color: #6fbef3;
-      }
-
-      .home-menu li a:hover,
-      .home-menu li a:focus {
-        background: none;
-        border: none;
-        color: #aecfe5;
-      }
-
-      #theme-picker {
-        margin: 1em;
-      }
-    </style>
+    <!-- Local Scripts -->
+    <script type="module" defer src="./dist/header.js"></script>
+    <script type="module" defer src="./dist/index.js"></script>
   </head>
   <body>
-    <header>
-      <div class="home-menu pure-menu pure-menu-horizontal">
-        <a class="pure-menu-heading" href=""> Themes for CodeMirror 6 Demo </a>
+    <header class="header">
+      <div class="menu-horizontal">
+        <h1 class="menu-heading">@fsegurai/codemirror-themes</h1>
 
-        <ul class="pure-menu-list">
-          <li class="pure-menu-item pure-menu-selected">
-            <a
+        <span class="divider-header"></span>
+
+        <ul class="menu-list">
+          <li class="menu-item menu-selected">
+            <md-icon-button id="themeToggle">
+              <md-icon></md-icon>
+            </md-icon-button>
+          </li>
+          <li class="menu-item menu-selected">
+            <md-icon-button
+              id="github"
               href="https://github.com/fsegurai/codemirror-themes"
-              class="pure-menu-link"
+              target="_blank"
             >
-              GitHub repo
-            </a>
+              <md-icon>
+                <svg
+                  class="custom-icon"
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 496 512"
+                >
+                  <path
+                    d="M165.9 397.4c0 2-2.3 3.6-5.2 3.6-3.3.3-5.6-1.3-5.6-3.6 0-2 2.3-3.6 5.2-3.6 3-.3 5.6 1.3 5.6 3.6m-31.1-4.5c-.7 2 1.3 4.3 4.3 4.9 2.6 1 5.6 0 6.2-2s-1.3-4.3-4.3-5.2c-2.6-.7-5.5.3-6.2 2.3m44.2-1.7c-2.9.7-4.9 2.6-4.6 4.9.3 2 2.9 3.3 5.9 2.6 2.9-.7 4.9-2.6 4.6-4.6-.3-1.9-3-3.2-5.9-2.9M244.8 8C106.1 8 0 113.3 0 252c0 110.9 69.8 205.8 169.5 239.2 12.8 2.3 17.3-5.6 17.3-12.1 0-6.2-.3-40.4-.3-61.4 0 0-70 15-84.7-29.8 0 0-11.4-29.1-27.8-36.6 0 0-22.9-15.7 1.6-15.4 0 0 24.9 2 38.6 25.8 21.9 38.6 58.6 27.5 72.9 20.9 2.3-16 8.8-27.1 16-33.7-55.9-6.2-112.3-14.3-112.3-110.5 0-27.5 7.6-41.3 23.6-58.9-2.6-6.5-11.1-33.3 2.6-67.9 20.9-6.5 69 27 69 27 20-5.6 41.5-8.5 62.8-8.5s42.8 2.9 62.8 8.5c0 0 48.1-33.6 69-27 13.7 34.7 5.2 61.4 2.6 67.9 16 17.7 25.8 31.5 25.8 58.9 0 96.5-58.9 104.2-114.8 110.5 9.2 7.9 17 22.9 17 46.4 0 33.7-.3 75.4-.3 83.6 0 6.5 4.6 14.4 17.3 12.1C428.2 457.8 496 362.9 496 252 496 113.3 383.5 8 244.8 8M97.2 352.9c-1.3 1-1 3.3.7 5.2 1.6 1.6 3.9 2.3 5.2 1 1.3-1 1-3.3-.7-5.2-1.6-1.6-3.9-2.3-5.2-1m-10.8-8.1c-.7 1.3.3 2.9 2.3 3.9 1.6 1 3.6.7 4.3-.7.7-1.3-.3-2.9-2.3-3.9-2-.6-3.6-.3-4.3.7m32.4 35.6c-1.6 1.3-1 4.3 1.3 6.2 2.3 2.3 5.2 2.6 6.5 1 1.3-1.3.7-4.3-1.3-6.2-2.2-2.3-5.2-2.6-6.5-1m-11.4-14.7c-1.6 1-1.6 3.6 0 5.9s4.3 3.3 5.6 2.3c1.6-1.3 1.6-3.9 0-6.2-1.4-2.3-4-3.3-5.6-2"
+                  />
+                </svg>
+              </md-icon>
+            </md-icon-button>
           </li>
         </ul>
       </div>
-      <div id="theme-picker">
-        <label for="theme-list">Theme:</label>
-        <select id="theme-list" name="themes"></select>
-      </div>
     </header>
 
-    <main>
-      <div id="codemirror"></div>
+    <nav id="tabLinkContainer" class="tab-link-container">
+      <div id="tabLinkContent" class="tab-link-content">
+        <!-- Routes will be injected here -->
+      </div>
+
+      <md-icon-button id="scrollUpButton" class="scroll-up-button">
+        <md-icon>keyboard_arrow_up</md-icon>
+      </md-icon-button>
+    </nav>
+
+    <main class="main-index">
+      <div id="markdown-body" class="markdown-body"></div>
     </main>
   </body>
 </html>
