+++ demo/playground.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <title>&#64;fsegurai/codemirror-theme | Demo</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" type="image/x-icon" href="favicon.ico" />
+    <link rel="icon" type="image/x-icon" href="public/favicon.ico" />
 
     <!-- Google Fonts -->
     <link rel="preconnect" href="https://fonts.googleapis.com" />
@@ -34,8 +34,8 @@
     </script>
 
     <!-- Local Scripts -->
-    <script type="module" async src="./dist/header.js"></script>
-    <script type="module" async src="./dist/playground.js"></script>
+    <script type="module" defer src="./dist/header.js"></script>
+    <script type="module" defer src="./dist/playground.js"></script>
   </head>
   <body>
     <header class="header">
