+++ demo/scripts/playground.ts
@@ -2,8 +2,8 @@ import { basicSetup, EditorView } from 'codemirror';
 import { markdown, markdownLanguage } from '@codemirror/lang-markdown';
 import { languages } from '@codemirror/language-data';
 import { Compartment } from '@codemirror/state';
-import mdSample from '../markdown.example';
-import themes from '../themes';
+import mdSample from './markdown.example';
+import themes from './themes';
 
 const elCM = document.querySelector('#codemirror')!;
 const themeConfig = new Compartment();
@@ -38,7 +38,8 @@ if (elList) {
       themeItem.textContent = themes[i].name;
       if (i === 0) {
         // Update the default option
-        defaultOption.querySelector('div[slot="headline"]')!.textContent = themes[i].name;
+        defaultOption.querySelector('div[slot="headline"]')!.textContent =
+          themes[i].name;
       } else {
         elItem.appendChild(themeItem);
       }
