+++ demo/scripts/themes.ts
@@ -1,26 +1,26 @@
-import { abcdef } from '../packages/abcdef/src';
-import { abyss } from '../packages/abyss/src';
-import { androidStudio } from '../packages/android-studio/src';
-import { andromeda } from '../packages/andromeda/src';
-import { basicDark } from '../packages/basic-dark/src';
-import { basicLight } from '../packages/basic-light/src';
-import { forest } from '../packages/forest/src';
-import { githubDark } from '../packages/github-dark/src';
-import { githubLight } from '../packages/github-light/src';
-import { gruvboxDark } from '../packages/gruvbox-dark/src';
-import { gruvboxLight } from '../packages/gruvbox-light/src';
-import { materialDark } from '../packages/material-dark/src';
-import { materialLight } from '../packages/material-light/src';
-import { monokai } from '../packages/monokai/src';
-import { nord } from '../packages/nord/src';
-import { palenight } from '../packages/palenight/src';
-import { solarizedDark } from '../packages/solarized-dark/src';
-import { solarizedLight } from '../packages/solarized-light/src';
-import { tokyoNightDay } from '../packages/tokyo-night-day/src';
-import { tokyoNightStorm } from '../packages/tokyo-night-storm/src';
-import { volcano } from '../packages/volcano/src';
-import { vsCodeDark } from '../packages/vscode-dark/src';
-import { vsCodeLight } from '../packages/vscode-light/src';
+import { abcdef } from '../../packages/abcdef/src';
+import { abyss } from '../../packages/abyss/src';
+import { androidStudio } from '../../packages/android-studio/src';
+import { andromeda } from '../../packages/andromeda/src';
+import { basicDark } from '../../packages/basic-dark/src';
+import { basicLight } from '../../packages/basic-light/src';
+import { forest } from '../../packages/forest/src';
+import { githubDark } from '../../packages/github-dark/src';
+import { githubLight } from '../../packages/github-light/src';
+import { gruvboxDark } from '../../packages/gruvbox-dark/src';
+import { gruvboxLight } from '../../packages/gruvbox-light/src';
+import { materialDark } from '../../packages/material-dark/src';
+import { materialLight } from '../../packages/material-light/src';
+import { monokai } from '../../packages/monokai/src';
+import { nord } from '../../packages/nord/src';
+import { palenight } from '../../packages/palenight/src';
+import { solarizedDark } from '../../packages/solarized-dark/src';
+import { solarizedLight } from '../../packages/solarized-light/src';
+import { tokyoNightDay } from '../../packages/tokyo-night-day/src';
+import { tokyoNightStorm } from '../../packages/tokyo-night-storm/src';
+import { volcano } from '../../packages/volcano/src';
+import { vsCodeDark } from '../../packages/vscode-dark/src';
+import { vsCodeLight } from '../../packages/vscode-light/src';
 
 const themes = [
   {
