+++ package.json
@@ -48,7 +48,8 @@
     "@codemirror/legacy-modes": "^6.4.2",
     "@lezer/markdown": "^1.3.2",
     "@material/web": "^2.2.0",
-    "codemirror": "^6.0.1"
+    "codemirror": "^6.0.1",
+    "marked": "^15.0.0"
   },
   "devDependencies": {
     "@codemirror/buildhelper": "^1.0.2",
